vbo-unstable merged into HEAD. A couple of minor changes in the process - water materials 6 and 8 switched back since GL doesn't seem to like starting on what is now 6; but is quite happy changing to it later. VC6 project file for MeshUpgrader added. VC6 ReferenceAppLayer output folder corrected.

+++ trunk/OgreMain/include/OgreVertexBoneAssignment.h
@@ -40,7 +40,7 @@ namespace Ogre
     */
     typedef struct VertexBoneAssignment_s
     {
-        unsigned short vertexIndex;
+        unsigned int vertexIndex;
         unsigned short boneIndex;
         Real weight;
